requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-telegram-bot==20.3
aiohttp==3.9.1
pyyaml==6.0.1
elasticsearch==8.11.0
prometheus-client==0.19.0
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
ansible-runner==2.3.3
huggingface-hub==0.19.4
sentence-transformers==2.2.2
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
